requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
PyNaCl==1.5.0
wavelink==3.2.0
aiohttp>=3.8.0
yarl>=1.9.2 